+++ Vault-Coder/test/utopia/coder/vault/test/MergeAccessTest.scala
@@ -1,0 +1,11 @@
+package utopia.coder.vault.test
+
+/**
+ * Merges a generated access point with an example
+ * @author Mikko Hilpinen
+ * @since 17.07.2025, v1.13
+ */
+object MergeAccessTest extends App
+{
+	
+}